Partially fix the long conversion

+++ src/CSnakes.Runtime/CPython/Long.cs
@@ -10,7 +10,7 @@ internal unsafe partial class CPythonAPI
     internal static partial nint PyLong_FromLong(long v);
 
     [LibraryImport(PythonLibraryName)]
-    internal static partial nint PyLong_AsLong(nint p);
+    internal static partial long PyLong_AsLong(nint p);
 
     internal static bool IsPyLong(nint p)
     {
